Set parser.prog when calling a method
This allows for a better automatically generated usage method when
calling a subcommand for CLIs that use objects.

As an example, prior to this commit, the `example/cli_object.py`
example would do the following:

    $ python examples/cli_object.py pull --help
    usage: cli_object.py [-h] [-b BRANCH]

    optional arguments:
      -h, --help            show this help message and exit
      -b BRANCH, --branch BRANCH

Two issues here are that the name set in the @hug.object decorator
is not respected, nor is the subcommand used.

After this commit, the example does the following:

    $ python examples/cli_object.py pull --help
    usage: git pull [-h] [-b BRANCH]

    optional arguments:
      -h, --help            show this help message and exit
      -b BRANCH, --branch BRANCH

The usage message now shows `git pull` as would be expected.